Initial commit
Previous commits contained sensitive data (all of my text messages). It
was useful to have this for testing and to be able to see what the
output looked like between commits. This is no longer necessary (I have
made realistic but fake data), so to avoid pushing anything that could
be used to retrieve my commits, I'm restarting the commit history.